requirements.txt
Django==3.1.3
django-macaddress
django-import-export
gunicorn==20.0.4

